make sure -lm is in $LIBS for the finite/isfinite test, fixes issues on
2002-08-03  Manish Singh  <yosh@gimp.org>

        * configure.in: make sure -lm is in $LIBS for the finite/isfinite
        test, fixes issues on newer FreeBSD

        * plug-ins/common/jpeg.c: cast statbuf.st_size to glong for
        g_snprintf

+++ plug-ins/jpeg/jpeg-load.c
@@ -1130,7 +1130,7 @@ background_jpeg_save (PreviewPersistent *pp)
 	  stat (pp->file_name, &buf);
 	  g_snprintf (temp, sizeof (temp),
 		      _("Size: %ld bytes (%02.01f kB)"),
-		      (gint) buf.st_size, (gdouble) (buf.st_size) / 1024.0);
+		      (glong) buf.st_size, (gdouble) (buf.st_size) / 1024.0);
 	  gtk_label_set_text (GTK_LABEL (preview_size), temp);
 	
 	  /* and load the preview */
